+++ src/main/java/com/mtcoding/minigram/stories/StoryResponse.java
@@ -4,7 +4,6 @@ import com.mtcoding.minigram.users.User;
 import lombok.Data;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 public class StoryResponse {
 
@@ -55,14 +54,6 @@ public class StoryResponse {
         }
     }
 
-    @Data
-    public static class ListDTO {
-        private List<DetailDTO> storyList;
-
-        public ListDTO(List<DetailDTO> storyList) {
-            this.storyList = storyList;
-        }
-    }
 
     @Data
     public static class DTO {
@@ -82,52 +73,6 @@ public class StoryResponse {
             this.status = story.getStatus();
             this.createdAt = story.getCreatedAt();
             this.updatedAt = story.getUpdatedAt();
-        }
-    }
-
-    @Data
-    public static class FeedDTO {
-        private Integer current;     // 현재 페이지(0-base)
-        private Integer size;        // 페이지당 개수
-        private Integer totalCount;  // 전체 글 수
-        private Integer totalPage;   // 전체 페이지 수
-        private Integer prev;        // current - 1
-        private Integer next;        // current + 1
-        private Boolean isFirst;     // current == 0
-        private Boolean isLast;      // (totalPage - 1) == current
-        private List<ItemDTO> storyHeadList;
-
-        public FeedDTO(List<ItemDTO> storyHeadList, Integer current, Integer totalCount) {
-            this.storyHeadList = storyHeadList;
-            this.current = current;
-            this.size = 10;
-            this.totalCount = totalCount;
-            this.totalPage = makeTotalPage(totalCount, size);
-            this.prev = Math.max(0, current - 1);
-            this.next = totalPage == 0 ? 0 : Math.min(totalPage - 1, current + 1);
-            this.isFirst = current == 0;
-            this.isLast = (totalPage - 1) == current;
-        }
-
-        private int makeTotalPage(int totalCount, int size) {
-            if (size <= 0) return 0;
-            int rest = (totalCount % size) > 0 ? 1 : 0;
-            return (totalCount / size) + rest;
-        }
-    }
-
-    @Data
-    public static class ItemDTO {
-        private Integer userId;
-        private String username;
-        private String profileImageUrl;
-        private Boolean hasUnseen; // 최근 5개의 스토리 중 한 개라도 안읽은 스토리가 있는지 여부
-
-        public ItemDTO(Integer userId, String username, String profileImageUrl, Boolean hasUnseen) {
-            this.userId = userId;
-            this.username = username;
-            this.profileImageUrl = profileImageUrl;
-            this.hasUnseen = hasUnseen;
         }
     }
 }
