+++ src/main/java/com/mtcoding/minigram/stories/StoryService.java
@@ -8,8 +8,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-
 @Slf4j
 @RequiredArgsConstructor
 @Service
@@ -36,40 +34,6 @@ public class StoryService {
         );
     }
 
-    public StoryResponse.ListDTO findAllMyStories(Integer currentUserId) {
-        List<Object[]> obsList = storyRepository.findAllMyStories(currentUserId);
-
-        List<StoryResponse.DetailDTO> detailDTOList = obsList.stream().map(ob -> {
-            Story story = (Story) ob[0];
-            Boolean isFollowing = (Boolean) ob[1];
-            Integer likeCount = ((Long) ob[2]).intValue();
-            Boolean isLiked = (Boolean) ob[3];
-
-            boolean isOwner = true;
-
-            return new StoryResponse.DetailDTO(story, isFollowing, isOwner, isLiked, likeCount);
-        }).toList();
-
-        return new StoryResponse.ListDTO(detailDTOList);
-    }
-
-    public StoryResponse.ListDTO findAllByUserId(Integer userId, Integer currentUserId) {
-        List<Object[]> obsList = storyRepository.findAllByUserId(userId, currentUserId);
-
-        List<StoryResponse.DetailDTO> detailDTOList = obsList.stream().map(ob -> {
-            Story story = (Story) ob[0];
-            Boolean isFollowing = (Boolean) ob[1];
-            Integer likeCount = ((Long) ob[2]).intValue();
-            Boolean isLiked = (Boolean) ob[3];
-
-            Boolean isOwner = story.getUser().getId().equals(currentUserId);
-
-            return new StoryResponse.DetailDTO(story, isFollowing, isOwner, isLiked, likeCount);
-        }).toList();
-
-        return new StoryResponse.ListDTO(detailDTOList);
-    }
-
     @Transactional
     public StoryResponse.DTO create(StoryRequest.CreateDTO reqDTO, User user) {
         Story story = reqDTO.toEntity(user);
@@ -94,25 +58,5 @@ public class StoryService {
         storyPS.delete();
 
         return new StoryResponse.DTO(storyPS);
-    }
-
-    public StoryResponse.FeedDTO getFeedStories(Integer page, Integer currentUserId) {
-        // 1. userId, username, profileImageUrl, hasUnseen 조회
-        List<Object[]> obsList = storyRepository.findFromFollowees(page, currentUserId);
-
-        // 2. ItemDTO 조립
-        List<StoryResponse.ItemDTO> itemDTOList = obsList.stream()
-                .map(obs -> new StoryResponse.ItemDTO(
-                        (Integer) obs[0],
-                        (String) obs[1],
-                        (String) obs[2],
-                        (Boolean) obs[3]
-                ))
-                .toList();
-
-        // 3. totalCount 조회
-        Integer totalCount = Math.toIntExact(storyRepository.totalCountFromFollowees(currentUserId));
-
-        return new StoryResponse.FeedDTO(itemDTOList, page, totalCount);
     }
 }
