+++ src/main/java/com/mtcoding/minigram/users/UserRepository.java
@@ -1,5 +1,7 @@
 package com.mtcoding.minigram.users;
 
+import com.mtcoding.minigram.posts.PostStatus;
+import com.mtcoding.minigram.stories.StoryStatus;
 import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
@@ -44,5 +46,58 @@ public class UserRepository {
 
     public User getReferenceById(Integer userId) {
         return em.getReference(User.class, userId);
+    }
+
+    public Optional<UserResponse.ProfileDTO> findProfileById(Integer profileUserId, Integer currentUserId, Boolean isOwner) {
+
+        try {
+            Object[] obs = em.createQuery("""
+                            SELECT u,
+                                   (SELECT COUNT(p.id)
+                                                 FROM Post p
+                                                WHERE p.user = u
+                                                  AND (p.status = :postActive OR (:isOwner = true AND p.status = :postHidden))
+                                              ),
+                                   (SELECT COUNT(f1.id) FROM Follow f1 WHERE f1.followee = u),
+                                   (SELECT COUNT(f2.id) FROM Follow f2 WHERE f2.follower = u),
+                                   CASE WHEN EXISTS (
+                                       SELECT 1 FROM Follow f
+                                        WHERE f.follower.id = :currentUserId AND f.followee = u
+                                   ) THEN true ELSE false END,
+                                   CASE WHEN EXISTS (
+                                       SELECT 1 FROM Story s
+                                        WHERE s.user = u AND s.status = :storyActive
+                                          AND NOT EXISTS (
+                                              SELECT 1 FROM StoryView v
+                                               WHERE v.story = s AND v.user.id = :currentUserId
+                                          )
+                                   ) THEN true ELSE false END
+                            FROM User u
+                            WHERE u.id = :profileUserId
+                            """, Object[].class)
+                    .setParameter("profileUserId", profileUserId)
+                    .setParameter("currentUserId", currentUserId)
+                    .setParameter("isOwner", isOwner)
+                    .setParameter("postActive", PostStatus.ACTIVE)
+                    .setParameter("postHidden", PostStatus.HIDDEN)
+                    .setParameter("storyActive", StoryStatus.ACTIVE)
+                    .getSingleResult();
+
+            // 조회 결과 순서 : user, postCount, followerCount, followingCount, isFollowing, hasUnseen
+            UserResponse.ProfileDTO profileDTO = new UserResponse.ProfileDTO(
+                    (User) obs[0],
+                    (Boolean) obs[5],
+                    isOwner,
+                    (Boolean) obs[4],
+                    ((Long) obs[1]).intValue(),
+                    ((Long) obs[2]).intValue(),
+                    ((Long) obs[3]).intValue()
+            );
+
+            return Optional.of(profileDTO);
+
+        } catch (Exception e) {
+            return Optional.empty();
+        }
     }
 }
