+++ src/main/java/com/mtcoding/minigram/users/UserService.java
@@ -2,6 +2,7 @@ package com.mtcoding.minigram.users;
 
 import com.mtcoding.minigram._core.error.ex.ExceptionApi400;
 import com.mtcoding.minigram._core.error.ex.ExceptionApi401;
+import com.mtcoding.minigram._core.error.ex.ExceptionApi404;
 import com.mtcoding.minigram._core.util.JwtUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -41,11 +42,25 @@ public class UserService {
         User userPS = userRepository.findByEmail(loginDTO.getEmail())
                 .orElseThrow(() -> new ExceptionApi401("유저네임 혹은 비밀번호가 틀렸습니다"));
 
-        Boolean isSame = bCryptPasswordEncoder.matches(loginDTO.getPassword(), userPS.getPassword());
+        boolean isSame = bCryptPasswordEncoder.matches(loginDTO.getPassword(), userPS.getPassword());
         if (!isSame) throw new ExceptionApi401("유저네임 혹은 비밀번호가 틀렸습니다");
 
         String accessToken = JwtUtil.create(userPS);
 
         return new UserResponse.LoginDTO(accessToken, userPS);
+    }
+
+    public UserResponse.ProfileDTO getUserProfile(Integer userId, Integer currentUserId) {
+        // 1. 프로필 조회 대상의 user.id 결정
+        Integer profileUserId = (userId == null) ? currentUserId : userId;
+
+        // 2. 본인 여부 판단
+        Boolean isOwner = profileUserId.equals(currentUserId);
+
+        // 3. profile 조회
+        UserResponse.ProfileDTO profileDTO = userRepository.findProfileById(profileUserId, currentUserId, isOwner)
+                .orElseThrow(() -> new ExceptionApi404("존재하지 않는 유저입니다."));
+
+        return profileDTO;
     }
 }
